// webapp/src/app/Model/Resource.ts
export class Resource {
    
    
    createdBy: String;
    type: String;
    bedType: String;
    medicineName:String;
    typeofEquipment:String;
    medicalEquipmentName:String;
    city: String;
    avalabilityPlace: String;
    address: String;
    contactPersonName: String;
    contactMobileNumber: String;
    isVerified: String;
    verifiedBy: String;
}